app/modules/audit/schemas.py
from datetime import datetime
from typing import Optional
from uuid import UUID

from pydantic import BaseModel


class AuditLogResponse(BaseModel):
    id: UUID
    entity_type: str
    entity_id: UUID
    action: str
    user_id: Optional[UUID]
    ip_address: Optional[str]
    user_agent: Optional[str]
    old_values: Optional[dict]
    new_values: Optional[dict]
    created_at: datetime

    model_config = {"from_attributes": True}


class AuditFilters(BaseModel):
    entity_type: Optional[str] = None
    entity_id: Optional[UUID] = None
    action: Optional[str] = None
    user_id: Optional[UUID] = None
    from_date: Optional[datetime] = None
    to_date: Optional[datetime] = None
